+++ packages/brick_json_generators/lib/json_serialize.dart
@@ -75,6 +75,8 @@ mixin JsonSerialize<_Model extends Model, _Annotation extends FieldSerializable>
         }
         return '${SharedChecker.withoutNullability(field.type)}.values.indexOf($fieldValue)';
       }
+    } else if (checker.toJsonMethod != null) {
+      return '$fieldValue.toJson()';
     }
 
     return null;