+++ packages/brick_build/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+## 2.0.1
+
+* Add `fromJsonConstructor` and `toJsonMethod` to `SharedChecker`, supporting easy discovery of standard methods `factory fromJson` and `#toJson`.
+
 ## 2.0.0
 
 * **BREAKING CHANGE** `brick.g.dart` and `adapters` are now saved to a `brick/` folder. This replaces the `app` folder. Please rename `app` to `brick` to safely migrate.